tts/client-vab-bot/python/voice-announcement-bot-eii/requirements.txt
grpcio
grpcio-tools
numpy
protobuf<4.21.0
sounddevice
pyzmq
